fix type

+++ apps/backend/Shelf.Infrastructure/Entity/File.cs
@@ -14,8 +14,8 @@ public class File
     public required long FileSize { get; set; }
     public required string MimeType { get; set; }
 
-    public required double Width { get; set; }
-    public required double Height { get; set; }
+    public required int Width { get; set; }
+    public required int Height { get; set; }
 
     public int? DurationInSeconds { get; set; } = null;
     public string Codec { get; set; } = string.Empty;
